Add OpenGL graphics module and related CMakeLists
A new module 'graphics' with a specific OpenGL support has been added, including the necessary CMakeLists files for building. Main.cpp file for OpenGL-related functions is included in this new module. The overall project's CMakeLists file is updated to add this new subdirectory. Changes to the README are also included for clean formatting.