Implement Index Buffer and Update Libraries
Added an Index Buffer in the graphics engine and updated code dependencies to work with the new addition. Created and implemented a new shader for the index buffer. Also included the addition of a new game editor library into the project. The README is updated to reflect these changes accurately.